+++ src/loader.rs
@@ -3,66 +3,9 @@
 use std::collections::BTreeMap;
 
 use hashlink::LinkedHashMap;
-use saphyr_parser::{Event, MarkedEventReceiver, Marker, Parser, ScanError, TScalarStyle, Tag};
+use saphyr_parser::{Event, MarkedEventReceiver, Marker, ScanError, TScalarStyle, Tag};
 
 use crate::{Hash, Yaml};
-
-/// Load the given string as an array of YAML documents.
-///
-/// The `source` is interpreted as YAML documents and is parsed. Parsing succeeds if and only
-/// if all documents are parsed successfully. An error in a latter document prevents the former
-/// from being returned.
-///
-/// Most often, only one document is loaded in a YAML string. In this case, only the first element
-/// of the returned `Vec` will be used. Otherwise, each element in the `Vec` is a document:
-///
-/// ```
-/// use saphyr::{load_from_str, Yaml};
-///
-/// let docs = load_from_str(r#"
-/// First document
-/// ---
-/// - Second document
-/// "#).unwrap();
-/// let first_document = &docs[0]; // Select the first YAML document
-/// // The document is a string containing "First document".
-/// assert_eq!(*first_document, Yaml::String("First document".to_owned()));
-///
-/// let second_document = &docs[1]; // Select the second YAML document
-/// // The document is an array containing a single string, "Second document".
-/// assert_eq!(second_document[0], Yaml::String("Second document".to_owned()));
-/// ```
-///
-/// # Errors
-/// Returns `ScanError` when loading fails.
-pub fn load_from_str(source: &str) -> Result<Vec<Yaml>, ScanError> {
-    load_from_iter(source.chars())
-}
-
-/// Load the contents of the given iterator as an array of YAML documents.
-///
-/// See [`load_from_str`] for details.
-///
-/// # Errors
-/// Returns `ScanError` when loading fails.
-pub fn load_from_iter<I: Iterator<Item = char>>(source: I) -> Result<Vec<Yaml>, ScanError> {
-    let mut parser = Parser::new(source);
-    load_from_parser(&mut parser)
-}
-
-/// Load the contents from the specified [`Parser`] as an array of YAML documents.
-///
-/// See [`load_from_str`] for details.
-///
-/// # Errors
-/// Returns `ScanError` when loading fails.
-pub fn load_from_parser<I: Iterator<Item = char>>(
-    parser: &mut Parser<I>,
-) -> Result<Vec<Yaml>, ScanError> {
-    let mut loader = YamlLoader::default();
-    parser.load(&mut loader, true)?;
-    Ok(loader.docs)
-}
 
 /// Main structure for parsing YAML.
 ///
