+++ src/yaml.rs
@@ -5,8 +5,9 @@
 use std::{convert::TryFrom, ops::Index, ops::IndexMut};
 
 use hashlink::LinkedHashMap;
+use saphyr_parser::{Parser, ScanError};
 
-use crate::loader::parse_f64;
+use crate::{loader::parse_f64, YamlLoader};
 
 /// A YAML node is stored as this `Yaml` enumeration, which provides an easy way to
 /// access your YAML document.
@@ -57,6 +58,63 @@ pub type Array = Vec<Yaml>;
 pub type Hash = LinkedHashMap<Yaml, Yaml>;
 
 impl Yaml {
+    /// Load the given string as an array of YAML documents.
+    ///
+    /// The `source` is interpreted as YAML documents and is parsed. Parsing succeeds if and only
+    /// if all documents are parsed successfully. An error in a latter document prevents the former
+    /// from being returned.
+    ///
+    /// Most often, only one document is loaded in a YAML string. In this case, only the first element
+    /// of the returned `Vec` will be used. Otherwise, each element in the `Vec` is a document:
+    ///
+    /// ```
+    /// use saphyr::Yaml;
+    ///
+    /// let docs = Yaml::load_from_str(r#"
+    /// First document
+    /// ---
+    /// - Second document
+    /// "#).unwrap();
+    /// let first_document = &docs[0]; // Select the first YAML document
+    /// // The document is a string containing "First document".
+    /// assert_eq!(*first_document, Yaml::String("First document".to_owned()));
+    ///
+    /// let second_document = &docs[1]; // Select the second YAML document
+    /// // The document is an array containing a single string, "Second document".
+    /// assert_eq!(second_document[0], Yaml::String("Second document".to_owned()));
+    /// ```
+    ///
+    /// # Errors
+    /// Returns `ScanError` when loading fails.
+    pub fn load_from_str(source: &str) -> Result<Vec<Self>, ScanError> {
+        Self::load_from_iter(source.chars())
+    }
+
+    /// Load the contents of the given iterator as an array of YAML documents.
+    ///
+    /// See [`Self::load_from_str`] for details.
+    ///
+    /// # Errors
+    /// Returns `ScanError` when loading fails.
+    pub fn load_from_iter<I: Iterator<Item = char>>(source: I) -> Result<Vec<Yaml>, ScanError> {
+        let mut parser = Parser::new(source);
+        Self::load_from_parser(&mut parser)
+    }
+
+    /// Load the contents from the specified [`Parser`] as an array of YAML documents.
+    ///
+    /// See [`Self::load_from_str`] for details.
+    ///
+    /// # Errors
+    /// Returns `ScanError` when loading fails.
+    pub fn load_from_parser<I: Iterator<Item = char>>(
+        parser: &mut Parser<I>,
+    ) -> Result<Vec<Yaml>, ScanError> {
+        let mut loader = YamlLoader::default();
+        parser.load(&mut loader, true)?;
+        Ok(loader.into_documents())
+    }
+
     define_as!(as_bool, bool, Boolean);
     define_as!(as_i64, i64, Integer);
