+++ tests/quickcheck.rs
@@ -3,7 +3,7 @@ extern crate quickcheck;
 
 use quickcheck::TestResult;
 
-use saphyr::{load_from_str, Yaml, YamlEmitter};
+use saphyr::{Yaml, YamlEmitter};
 
 quickcheck! {
     fn test_check_weird_keys(xs: Vec<String>) -> TestResult {
@@ -13,7 +13,7 @@ quickcheck! {
             let mut emitter = YamlEmitter::new(&mut out_str);
             emitter.dump(&input).unwrap();
         }
-        match load_from_str(&out_str) {
+        match Yaml::load_from_str(&out_str) {
             Ok(output) => TestResult::from_bool(output.len() == 1 && input == output[0]),
             Err(err) => TestResult::error(err.to_string()),
         }
